server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
redis==5.0.8
celery==5.4.0
stripe==10.11.0
scenedetect==0.6.4
opencv-python==4.10.0.84
mediapipe==0.10.14
openai-whisper==20240930
pytest==8.3.3
